PostureAnalyzer/requirements.txt
streamlit
mediapipe
opencv-python
numpy
google-genai
plotly
Pillow
